jwt update

+++ src/routes/wine/deleteAdmin.js
@@ -1,18 +1,22 @@
 const { prisma } = require("../../prisma");
 
 async function deleteWineAdmin(req, res) {
-  try {
-    await prisma.wine.delete({
-      where: {
-        id: req.body.id,
-      },
-    });
+  if (req.user.admin == true) {
+    try {
+      await prisma.wine.delete({
+        where: {
+          id: req.body.id,
+        },
+      });
 
-    res.status(200).send("deleted wine");
-  } catch (error) {
-    console.log(error);
+      res.status(200).send("deleted wine");
+    } catch (error) {
+      console.log(error);
 
-    res.status(400).send("erreur");
+      res.status(400).send("erreur");
+    }
+  } else {
+    res.status(403).send("Accès refusé, vous n'êtes pas administrateur");
   }
 }
